vtk gui app update
